refactor: change packages folder name as lib in the apps
